+++ source/Lib/CommonLib/RdCostWeightPrediction.cpp
@@ -41,12 +41,12 @@
 
 #include <math.h>
 
-static Distortion xCalcHADs2x2w(const WPScalingParam &wpCur, const Pel *piOrg, const Pel *piCurr, int strideOrg,
-                                int strideCur, int step);
-static Distortion xCalcHADs4x4w(const WPScalingParam &wpCur, const Pel *piOrg, const Pel *piCurr, int strideOrg,
-                                int strideCur, int step);
-static Distortion xCalcHADs8x8w(const WPScalingParam &wpCur, const Pel *piOrg, const Pel *piCurr, int strideOrg,
-                                int strideCur, int step);
+static Distortion xCalcHADs2x2w(const WPScalingParam &wpCur, const Pel *piOrg, const Pel *piCurr, ptrdiff_t strideOrg,
+                                ptrdiff_t strideCur, int step);
+static Distortion xCalcHADs4x4w(const WPScalingParam &wpCur, const Pel *piOrg, const Pel *piCurr, ptrdiff_t strideOrg,
+                                ptrdiff_t strideCur, int step);
+static Distortion xCalcHADs8x8w(const WPScalingParam &wpCur, const Pel *piOrg, const Pel *piCurr, ptrdiff_t strideOrg,
+                                ptrdiff_t strideCur, int step);
 
 // --------------------------------------------------------------------------------------------------------------------
 // SAD
@@ -60,8 +60,8 @@ Distortion RdCostWeightPrediction::xGetSADw( const DistParam &rcDtParam )
   const Pel            *piOrg      = rcDtParam.org.buf;
   const Pel            *piCur      = rcDtParam.cur.buf;
   const int             cols       = rcDtParam.org.width;
-  const int             strideCur  = rcDtParam.cur.stride;
-  const int             strideOrg  = rcDtParam.org.stride;
+  const ptrdiff_t       strideCur  = rcDtParam.cur.stride;
+  const ptrdiff_t       strideOrg  = rcDtParam.org.stride;
   const ComponentID     compID     = rcDtParam.compID;
 
   CHECK( compID >= MAX_NUM_COMPONENT, "Invalid component" );
@@ -223,8 +223,8 @@ Distortion RdCostWeightPrediction::xGetSSEw( const DistParam &rcDtParam )
   const Pel            *piOrg      = rcDtParam.org.buf;
   const Pel            *piCur      = rcDtParam.cur.buf;
   const int             cols       = rcDtParam.org.width;
-  const int             strideCur  = rcDtParam.cur.stride;
-  const int             strideOrg  = rcDtParam.org.stride;
+  const ptrdiff_t       strideCur  = rcDtParam.cur.stride;
+  const ptrdiff_t       strideOrg  = rcDtParam.org.stride;
   const ComponentID     compID     = rcDtParam.compID;
 
   CHECK( rcDtParam.subShift != 0, "Subshift not supported" ); // NOTE: what is this protecting?
@@ -282,8 +282,8 @@ Distortion RdCostWeightPrediction::xGetSSEw( const DistParam &rcDtParam )
 // HADAMARD with step (used in fractional search)
 // --------------------------------------------------------------------------------------------------------------------
 //! get weighted Hadamard cost for 2x2 block
-Distortion xCalcHADs2x2w(const WPScalingParam &wpCur, const Pel *piOrg, const Pel *piCur, int strideOrg, int strideCur,
-                         int step)
+Distortion xCalcHADs2x2w(const WPScalingParam &wpCur, const Pel *piOrg, const Pel *piCur, ptrdiff_t strideOrg,
+                         ptrdiff_t strideCur, int step)
 {
   const int round  = wpCur.round;
   const int shift  = wpCur.shift;
@@ -320,8 +320,8 @@ Distortion xCalcHADs2x2w(const WPScalingParam &wpCur, const Pel *piOrg, const Pe
 
 
 //! get weighted Hadamard cost for 4x4 block
-Distortion xCalcHADs4x4w(const WPScalingParam &wpCur, const Pel *piOrg, const Pel *piCur, int strideOrg, int strideCur,
-                         int step)
+Distortion xCalcHADs4x4w(const WPScalingParam &wpCur, const Pel *piOrg, const Pel *piCur, ptrdiff_t strideOrg,
+                         ptrdiff_t strideCur, int step)
 {
   const int round  = wpCur.round;
   const int shift  = wpCur.shift;
@@ -430,8 +430,8 @@ Distortion xCalcHADs4x4w(const WPScalingParam &wpCur, const Pel *piOrg, const Pe
 
 
 //! get weighted Hadamard cost for 8x8 block
-Distortion xCalcHADs8x8w(const WPScalingParam &wpCur, const Pel *piOrg, const Pel *piCur, int strideOrg, int strideCur,
-                         int step)
+Distortion xCalcHADs8x8w(const WPScalingParam &wpCur, const Pel *piOrg, const Pel *piCur, ptrdiff_t strideOrg,
+                         ptrdiff_t strideCur, int step)
 {
   Distortion sad=0;
   TCoeff diff[64], m1[8][8], m2[8][8], m3[8][8];
@@ -555,8 +555,8 @@ Distortion RdCostWeightPrediction::xGetHADsw( const DistParam &rcDtParam )
   const Pel        *piCur      = rcDtParam.cur.buf;
   const int         rows       = rcDtParam.org.height;
   const int         cols       = rcDtParam.org.width;
-  const int         strideCur  = rcDtParam.cur.stride;
-  const int         strideOrg  = rcDtParam.org.stride;
+  const ptrdiff_t   strideCur  = rcDtParam.cur.stride;
+  const ptrdiff_t   strideOrg  = rcDtParam.org.stride;
   const int         step       = rcDtParam.step;
   const ComponentID compIdx    = rcDtParam.compID;
   CHECK(compIdx>=MAX_NUM_COMPONENT, "Invalid component");
@@ -566,8 +566,8 @@ Distortion RdCostWeightPrediction::xGetHADsw( const DistParam &rcDtParam )
 
   if ((rows % 8 == 0) && (cols % 8 == 0))
   {
-    const int offsetOrg = strideOrg << 3;
-    const int offsetCur = strideCur << 3;
+    const ptrdiff_t offsetOrg = strideOrg << 3;
+    const ptrdiff_t offsetCur = strideCur << 3;
     for (int y = 0; y < rows; y += 8)
     {
       for (int x = 0; x < cols; x += 8)
@@ -580,8 +580,8 @@ Distortion RdCostWeightPrediction::xGetHADsw( const DistParam &rcDtParam )
   }
   else if ((rows % 4 == 0) && (cols % 4 == 0))
   {
-    const int offsetOrg = strideOrg << 2;
-    const int offsetCur = strideCur << 2;
+    const ptrdiff_t offsetOrg = strideOrg << 2;
+    const ptrdiff_t offsetCur = strideCur << 2;
 
     for (int y = 0; y < rows; y += 4)
     {
